update outdated info
updated the part that said the ISO installs HTTPS and Wireguard by default but the script doesn't but will in the future.  The future is now and those packages are installed by the script as well.  I added a note about installing via script for root and sudo access for advanced users.